Moves DeviceDataManager's extension to its own file
